test_main.py: check error messages instead of exceptions types and attributes
Tests using checking the type and attributes of the exceptions look
like this:

   with pytest.raises(SomeError) as e:
      some_code()
   assert e.value.field == "expected value"

and they belong in low-level tests

In high-level tests such as `test_main()` we write assertions about
the error messages instead.

This makes sure there are no bugs in the implementation of the
`print_error()` method.

This matters because if there is a bug in a `print_error()`
method, users will get a confusing "double" stacktrace:
("During handling of the above exception, another exception occurred".)