Don't ask for current version when using `init`
Instead, we just parse the new version from the command line.

Note that we use docopt instead of argparse because we can't express
a "default subparser" with argparse.

We still need the 'if new_version=="init"' check though :/

Side-note: we also got rid of argparse Namespace object that was
used everywhere without any type checks