Relax dependency requirement on cli-ui
According to the [dependency specification](https://python-poetry.org/docs/dependency-specification/)
`^1.2.3` means `>=1.2.3 <2.0.0`, but `^0.2.3` means `>=0.2.3 <0.3.0`.
Your current version of cli-ui is `0.15.2` which is out of the scope of `^0.10.3` (which actually means `>=0.10.3 <0.11.0`).
In order to submit your `tbump` package to conda-forge, I need to submit `cli-ui` to conda-forge as well. The CI complains on this issue.